extra analysis tools and optimisation
the least and most sold products, daily profit graph, the highest revenue generating products were added to the analysis section. readme was created. css edits were made. when the product was sold, the active field was added to the product because it did not appear in the sales and the deleted products were set to be deactive. in this way, the products in the sales remained permanent.